pyproject.toml
[project]
name = "auto-gmail-creator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "appium-python-client>=5.2.4",
]
